Set up 2 page create report dialog
