+++ apps/i18n/common/en_us.json
@@ -348,7 +348,9 @@
   "createSectionsInfo": "Create new sections and add students to them. Sections help you organize students into smaller groups so you can track their progress and manage their accounts.",
   "createReport": "Create report",
   "createStandardsReport": "Create a CSTA Standards report",
+  "createStandardsReportPage1Body": "The report will be generated as a PDF so you easily print or email it.",
   "createStandardsReportSampleNoteText": "This note will appear in the report.",
+  "createStandardsReportStep1": "Step 1:",
   "createStandardsReportStep2": "Step 2: Add a personalized note to your report",
   "createStandardsReportPrompt": "Say more about what’s happening in your classroom. ",
   "createStandardsReportSuggestion": "**Suggestions:**",
@@ -964,6 +966,7 @@
   "newSectionEllipsis": "New section...",
   "newSectionAdd": "Add a new classroom section",
   "newSectionCreate": "Create a new section",
+  "next": "Next",
   "nextLevel": "Congratulations! You completed Puzzle {puzzleNumber}.",
   "nextLevelTrophies": "Congratulations! You completed Puzzle {puzzleNumber} and won {numTrophies, plural, one {a trophy} other {# trophies}}.",
   "nextPuzzle": "Next puzzle",
