+++ apps/src/templates/sectionProgress/standards/LessonStatusList.story.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import {UnconnectedLessonStatusList as LessonStatusList} from './LessonStatusList';
+
+export default storybook => {
+  return storybook
+    .storiesOf('Standards/LessonStatusList', module)
+    .add('overview', () => {
+      return <LessonStatusList />;
+    });
+};
